integration of the new API for polynomial matrix
