+++ linbox/matrix/densematrix/blas-submatrix.inl
@@ -119,12 +119,11 @@ namespace LinBox {
         _field(M.field())
     {
         //std::cout<<"BlasSubMatrix Copy constructor 2: "<<getPointer()<<" ="<<M.getPointer()<<"--->"<<&M<<std::endl;
-        // std::cout<<M<<std::Endl;
+        // std::cout<<M<<std::endl;
         // std::cout<<*this<<std::endl;
     }
 
-    // this constructor make compilation issues (anyway it seems that it is useless as constructing a blassub<non const> from a blassub<const> is not allowed) 
-#if 0
+
     template < class _Matrix >
     BlasSubmatrix<_Matrix>::BlasSubmatrix(const nonconstSubMatrixType &M) :
         _ptr(M.getPointer()),
@@ -137,16 +136,17 @@ namespace LinBox {
         // std::cout<<M<<std::endl;
         // std::cout<<*this<<std::endl;
     }
-#endif    
+
     
     template < class _Matrix >
-    BlasSubmatrix<_Matrix>& BlasSubmatrix<_Matrix>::copy (const BlasSubmatrix<_Matrix> & M){
+    template < class _AnyMatrix >
+    BlasSubmatrix<_Matrix>& BlasSubmatrix<_Matrix>::copy (const _AnyMatrix & M){
         //std::cout<<"BlasSubMatrix Copy Method\n";
         if (_row == M.rowdim() && _col == M.coldim()){
             // for (size_t i=0;i<_row;i++)
             //     for(size_t j=0;j<_col;j++)
             //         setEntry(i,j,M.getEntry(i,j));
-            FFLAS::fassign(_field, _row, _col, M._ptr, M._stride , _ptr, _stride); 
+            FFLAS::fassign(_field, _row, _col, M.getPointer(), M.getStride() , _ptr, _stride); 
         }
         else
             throw LinBoxError("Calling copy from BlasSubMatrix with matrices of different dimension ... not allowed");
