+++ linbox/matrix/polynomial-matrix.h
@@ -1,9 +1,7 @@
 /*
- * Copyright (C) 2013  Pascal Giorgi
- *                     Romain Lebreton
+ * Copyright (C) 2020  Pascal Giorgi
  *
  * Written by Pascal Giorgi <pascal.giorgi@lirmm.fr>
- *            Romain Lebreton <lebreton@lirmm.fr>
  *
  * ========LICENCE========
  * This file is part of the library LinBox.
@@ -23,16 +21,13 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  * ========LICENCE========
  */
-#ifndef __OLD_POLYNOMIAL_MATRIX
-#include "linbox/matrix/new-polynomial-matrix.h"
-#else
-
 #ifndef __LINBOX_polynomial_matrix_H
 #define __LINBOX_polynomial_matrix_H
 #include <vector>
-#include "linbox/vector/subvector.h"
+#include "linbox/vector/vector.h"
 #include "linbox/matrix/dense-matrix.h"
 #include "linbox/matrix/matrix-domain.h"
+
 #include "linbox/field/hom.h"
 #include "fflas-ffpack/utils/align-allocator.h"
 #include "givaro/modular.h"
@@ -49,176 +44,170 @@ uint64_t max_memory=0, cur_memory=0;
 #define DEL_MEM(X) ;     
 #endif
 
+
+
 #define COPY_BLOCKSIZE 32
 
 namespace LinBox{
 
-	enum PMType {polfirst, matfirst};
-	enum PMStorage {plain, view, const_view};
+    
+	enum PMType {polfirst, matfirst, matrowfirst};
 
-	// Generic handler class for Polynomial Matrix
-	template<size_t type, size_t storage, class Field>
-	class PolynomialMatrix;
 
-	template<typename Field> uint64_t element_storage(const Field& F)      { integer p;F.characteristic(p); return length(p);}
+    // matrix example
+    /*  [ 1+2X+3X^2  4+5X+6X^2 ]
+     *  [ 7+8X+9X^2  10+11X+12X^2 ]
+     */
+    
+    /* polfirst storage : raw matrix of dimension (_row*_col) x _size
+     *  [  1  2  3 ]
+     *  [  4  5  6 ]  -> polynomial are stored contiguously (good for FFT)
+     *  [  7  8  9 ]  -> storage is not BLAS compliant
+     *  [ 10 11 12 ]
+     *
+     * matfirst storage : raw matrix of dimension _size x (_row*_col) (basically transposed of polfirst)
+     *  [ 1 4 7 10 ]
+     *  [ 2 5 8 11 ]  -> storage is BLAS compliant
+     *  [ 3 6 9 12 ]  -> polynomial are not stored contiguously (gap of row*col, not good for FFT)
+     *
+     * matrowfirst storage: raw matrix of dimension _row x (_size*_col) 
+     * [ 1  4 2  5 3  6 ]  -> storage is BLAS compliant (using stride)
+     * [ 7 10 8 11 9 12 ]  -> polynomial are not stored contiguously (gap of row, might be good for FFT)
+     *
+     */
+     
+    // PG: the following does not make sense with characteristic 0 field !!! 
+	template<typename Field> uint64_t element_storage(const Field& F)      {
+        integer p;F.characteristic(p);
+        if (p==0) throw LinboxError("PolynomialMatrix: use of element_storage with characterisrtic zero field is not allowed, aborting ...");
+        return length(p);}
 	template<> uint64_t element_storage(const Givaro::Modular<Givaro::Integer> &F) { integer p;F.characteristic(p); return length(p)+sizeof(Givaro::Integer);}
-	
+    
+    
+	// Generic handler class for Polynomial Matrix
+	template<class Field, enum PMType>
+	class PolynomialMatrix;
+    
+    template<class MatPoly>
+	class SubPolynomialMatrix;
+    
+    
 	// Class for Polynomial Matrix stored as a Matrix of Polynomials
 	template<class _Field>
-	class PolynomialMatrix<PMType::polfirst,PMStorage::plain,_Field> {
-	public:
-		typedef _Field Field;
-		typedef typename Field::Element   Element;
-		typedef BlasMatrix<Field>          Matrix;
-		//typedef typename std::vector<Element>::iterator  Iterator;
-		//typedef typename std::vector<Element>::const_iterator  ConstIterator;
+	class PolynomialMatrix<_Field, PMType::polfirst> {
+    public:
+        typedef _Field                         Field;
+		typedef typename Field::Element      Element;
 		typedef std::vector<Element,AlignedAllocator<Element, Alignment::DEFAULT>> VECT;
-		typedef typename VECT::iterator  Iterator;
-		typedef typename VECT::const_iterator  ConstIterator;
-		//typedef vector<Element>        Polynomial;
-		typedef Subvector<Iterator,ConstIterator>   Polynomial;
-		typedef PolynomialMatrix<PMType::polfirst,PMStorage::plain,_Field>  Self_t;
-		typedef PolynomialMatrix<PMType::matfirst,PMStorage::plain,_Field> Other_t;
+        typedef BlasVector<Field>                             Data;        
 
-		//PolynomialMatrix() {}
+        typedef BlasMatrix<Field>                                   Matrix;
+        typedef const BlasMatrix<Field>                        constMatrix;
+        typedef BlasSubvector<Data>                             Polynomial;
+        typedef const BlasSubvector<const Data>            constPolynomial;        
 
+		typedef PolynomialMatrix<Field, PMType::polfirst>  Self_t;
+        typedef SubPolynomialMatrix<Self_t>                  view;
+        typedef SubPolynomialMatrix<const Self_t>      const_view;
+        
 		// construct a polynomial matrix in f[x]^(m x n) of degree (s-1)
-		PolynomialMatrix(const Field& f, size_t r, size_t c, size_t s, size_t stor=0) :
-			_store((stor?stor:s)), _repview(r*c),_rep(r*c*_store,f.zero), _row(r), _col(c), _size(s), _fld(&f) {
-			for (size_t i=0;i<_row;i++)
-				for (size_t j=0;j<_col;j++)
-					_repview[i*_col+j]= Polynomial(_rep.begin()+(i*_col+j)*_store,_size);
-			//integer p;
-			//std::cout<<"MatrixP allocating : "<<r*c*s*length(f.characteristic(p))/1000000.<<"Mo"<<std::endl;
-			//std::cout<<"(ALLOC) PolynomialMatrix<polfirst> at "<<this<<" : "<<r<<"x"<<c<<" - size= "<<s<<" ==> "<<MB(realmeminfo())<<" Mo   "<<STR_MEMINFO<<std::endl;
-			ADD_MEM(realmeminfo());
-		}
+		PolynomialMatrix(const Field& f, size_t r, size_t c, size_t s)
+            : _rep(f,r*c*s), _row(r), _col(c), _size(s) {}
 
-		PolynomialMatrix(const Self_t&) = delete;
-		
-		~PolynomialMatrix(){
-			DEL_MEM(realmeminfo());
-			_rep.clear();
-			//std::cout<<"(FREE) PolynomialMatrix<polfirst> at "<<this<<" : "<<_row<<"x"<<_col<<" - size= "<<_store<<" ==> "<<MB(realmeminfo())<<" Mo   "<<STR_MEMINFO<<std::endl;
 
-			//integer p;
-			//std::cout<<"MatrixP Desallocating : "<<_row*_col*_store*length(_fld->characteristic(p))/1000000.<<"Mo"<<std::endl;
-			
-		}
-			
-		void clear(){
-			_rep.resize(0);
-		}
+        /*******************************
+         * Data access functionnalities*
+         ******************************/
+        
+        // retrieve the polynomial at entry (i,j) in the matrix
+        Polynomial       operator()(size_t i, size_t j)      {return operator()(i*_col+j);}
+        constPolynomial  operator()(size_t i, size_t j)const {return operator()(i*_col+j);}
 
-		// set the matrix A to the matrix of degree k in the polynomial matrix
-		void setMatrix(Matrix& A, size_t k) const {
-			typename Matrix::Iterator it=A.Begin();
+        // retrieve the polynomial at the position i in the storage of the matrix
+        Polynomial       operator()(size_t i)      {return      Polynomial(_rep, i*_size,1,_size);}
+        constPolynomial  operator()(size_t i)const {return constPolynomial(_rep, i*_size,1,_size);}
+
+		// get write access to the the k-th coeff  of the ith matrix entry
+        Element& ref(size_t i, size_t k)      {return _rep[i*_size+k];}
+        // get read access to the the k-th coeff  of the ith matrix entry
+        Element  get(size_t i, size_t k) const {return _rep[i*_size+k];}
+        // get write access to the the k-th coeff  of the entry (i,j) in matrix 
+        Element& ref(size_t i, size_t j, size_t k)      { return _rep[(i*_col+j)*_size+k];}
+        // get read access to the the k-th coeff  of the entry (i,j) in matrix 
+        Element  get(size_t i, size_t j, size_t k) const { return _rep[(i*_col+j)*_size+k];}
+
+		Element*       getPointer()       {return _rep.getPointer();}
+		const Element* getPointer() const {return _rep.getConstPointer();}
+
+        view       at(size_t i, size_t j)      {return view(*this,i,j);}
+        const_view at(size_t i, size_t j)const {return const_view(*this,i,j);}
+
+        // copy the matrix of degree k into A
+        template<typename DenseMatrix>
+		DenseMatrix& getMatrix(DenseMatrix& A, size_t k) const {            
+			auto it=A.Begin();
 			for(size_t i=0;i<_row*_col;i++,it++)
-				*it = get(i,k);			
+				*it = get(i,k);
+            return A;
 		}
-		// retrieve the matrix of degree k in the polynomial matrix
-		Matrix     operator[](size_t k)const {
+
+        // copy the matrix A into the matrix of degree
+        template<typename DenseMatrix>
+		void setMatrix(const DenseMatrix& A, size_t k)  {
+            //std::cout<<"SET MATRIX: "<<std::endl;
+			auto it=A.Begin();
+			for(size_t i=0;i<_row*_col;i++,it++){
+                //std::cout<<*it<<" ";
+                ref(i,k)=*it ;
+            }
+            //std::cout<<std::endl;
+		}
+
+        // retrieve the matrix of degree k in the polynomial matrix
+		Matrix     operator[](size_t k) const {
 			Matrix A(field(), _row, _col);
-			setMatrix(A,k);
+			getMatrix(A,k);
 			return A;
 		}
-		// retrieve the polynomial at entry (i,j) in the matrix
-		inline Polynomial&       operator()(size_t i, size_t j)      {return operator()(i*_col+j);}
-		inline const Polynomial& operator()(size_t i, size_t j)const {return operator()(i*_col+j);}
-		// retrieve the polynomial at the position i in the storage of the matrix
-		inline Polynomial&       operator()(size_t i)      {return _repview[i];}
-		inline const Polynomial& operator()(size_t i)const {return _repview[i];}
-
+        
 		// resize the polynomial length of the polynomial matrix
 		void resize(size_t s){
-			if (s==_store) return;
-			//std::cout<<"MATPOL RESIZING : "<<_store<<" --> "<<s<<std::endl;
-			if (s>_store){
-				_rep.resize(s*_row*_col);
+			if (s==_size) return;
+			if (s>_size){
+				_rep.resize(s*_row*_col, field().zero);
 				size_t k=s*_row*_col-1;
 				for(size_t i=0;i<_row*_col;i++){
-					size_t j=s;
-					for(;j>=_size;j--,k--)
-						_rep[k]=_fld->zero;
+					size_t j=_size;
 					for(;j>size_t(-1);j--,k--)
-						_rep[k]=_rep[i*_store+j];
+						_rep[k]=_rep[i*_size+j];
 				}
 			}
 			else {
 				size_t k=0;
 				for(size_t i=0;i<_row*_col;i++)
 					for (size_t j=0;j<s;j++,k++)
-						_rep[k]=_rep[i*_store+j];
+						_rep[k]=_rep[i*_size+j];
 				_rep.resize(s*_row*_col);
-				//_rep.shrink_to_fit();
 			}
-			integer p;_fld->characteristic(p); size_t bb=p.bitsize(); if(bb>64) bb+=128; bb/=8;
-
-#ifdef MEM_PMBASIS
-			size_t mem=realmeminfo();
-			ADD_MEM(realmeminfo());
-			DEL_MEM(mem);
-#endif			
-			_store=s;
-			setsize(s);
-		}
-
-		void changesize(size_t s){
-			if (s <=_store){			
-				for (size_t i=0;i<_row*_col;i++)
-					_repview[i]= Polynomial(_rep.begin()+i*_store,s);
-				_size=s;
-			}
-			else {
-				std::cerr<<"ERROR: in PolynomialMatrix<polfirs> -> setsize with too large size"<<std::endl;				
-				throw LinboxError("LinBox ERROR: PolynomialMatrix setsize \n");
-			}
-
-		}
-		void setsize(size_t s){
-			if (s <=_store){
-				for(size_t i=0;i<_row*_col;i++){
-					for(size_t j=_size ;j<s;j++)
-						_rep[i*_store+j]=_fld->zero;
-				}
-				for (size_t i=0;i<_row*_col;i++)
-					_repview[i]= Polynomial(_rep.begin()+i*_store,s);
-				_size=s;
-			}
-			else {
-				std::cerr<<"ERROR: in PolynomialMatrix<polfirs> -> setsize with too large size"<<std::endl;				
-				throw LinboxError("LinBox ERROR: PolynomialMatrix setsize \n");
-			}	
-		}
+			_size=s;
+		}	
 		
 		
 		// copy elt from M[beg..end], _size must be >= j-i
-		void copy(const Self_t& M, size_t beg, size_t end){
+		void copy(const PolynomialMatrix<Field, PMType::polfirst>& M, size_t beg, size_t end, size_t start=0){
 			//cout<<"copying.....polfirst to polfirst.....same field"<<endl;
 			for (size_t k=0;k<_row*_col;k++){
 				size_t j=0;
 				for (size_t i=beg;i<=end;i++){
-					ref(k,j)=M.get(k,i);
+					ref(k,start+j)=M.get(k,i);
 					j++;
 				}
 			}
 		}
-		template<typename OtherField>
-		void copy(const PolynomialMatrix<PMType::polfirst,PMStorage::plain,OtherField> & M, size_t beg, size_t end){
-			//cout<<"copying.....polfirst to polfirst.....other field"<<endl;
-			Hom<OtherField,Field> hom(M.field(),field()) ;
-			for (size_t k=0;k<_row*_col;k++){
-				size_t j=0;
-				for (size_t i=beg;i<=end;i++,j++){
-					hom.image(ref(k,j),M.get(k,i));
-				}
-			}
-		}
 
-		// copy elt from M[beg..end], _size must be >= end-beg+1
+        // copy elt from M[beg..end], _size must be >= end-beg+1
 		// M is stored as a Polynomial of Matrices
-		template <size_t storage>
-		void copy(const PolynomialMatrix<PMType::matfirst,storage,Field>& M, size_t beg, size_t end){
+		void copy(const PolynomialMatrix<Field, PMType::matfirst>& M, size_t beg, size_t end, size_t start=0){
 			//std::cout<<"copying.....matfirst to polfirst.....same field"<<std::endl;
 			const size_t ls = COPY_BLOCKSIZE;
 			for (size_t i = beg; i <= end; i+=ls)
@@ -226,13 +215,28 @@ namespace LinBox{
 					// Rk: the two loop must be interchanged in some cases
 					for (size_t _i = i; _i < std::min(end+1, i + ls); _i++)
 						for (size_t _j = j; _j < std::min(_col * _row, j + ls);++_j)
-							ref(_j,_i-beg)= M.get(_j,_i);
+							ref(_j,start+_i-beg)= M.get(_j,_i);
 		}
 
+        // copy elt from M[beg..end], _size must be >= end-beg+1
+		// M is stored with the hybrid format matrowfirst
+		void copy(const PolynomialMatrix<Field, PMType::matrowfirst>& M, size_t beg, size_t end, size_t start=0){
+			//std::cout<<"copying.....matrowfirst to polfirst.....same field"<<std::endl;
+			const size_t ls = COPY_BLOCKSIZE;
+            for(size_t k=0;k<_row;k++){
+                for (size_t i = beg; i <= end; i+=ls)
+                    for (size_t j = 0; j < _col; j+=ls)        
+                        for (size_t _i = i; _i < std::min(end+1, i + ls); _i++)
+                            for (size_t _j = j; _j < std::min(_col, j + ls);++_j)
+                                ref(k,_j,start+_i-beg)= M.get(k,_j,_i);
+            }
+        }
+
+        
 		// copy elt from M[beg..end], _size must be >= end-beg+1
 		// M is stored as a Polynomial of Matrices with a different field
-		template<size_t storage, typename OtherField>
-		void copy(const PolynomialMatrix<PMType::matfirst,storage,OtherField> & M, size_t beg, size_t end){
+		template<typename OtherField>
+		void copy(const PolynomialMatrix<OtherField, PMType::matfirst> & M, size_t beg, size_t end, size_t start=0){
 			//std::cout<<"copying.....matfirst to polfirst.....other field"<<std::endl;
 			const size_t ls = COPY_BLOCKSIZE;
 			Hom<OtherField,Field> hom(M.field(),field()) ;
@@ -240,9 +244,41 @@ namespace LinBox{
 				for (size_t j = 0; j < _col * _row; j+=ls)
 					for (size_t _i = i; _i < std::min(end+1, i + ls); _i++) {
 						for (size_t _j = j; _j < std::min(_col * _row, j + ls);++_j)
-							hom.image(ref(_j,_i-beg),M.get(_j,_i) );
+							hom.image(ref(_j,start+_i-beg),M.get(_j,_i) );
 					}
 		}
+
+
+
+        template<typename OtherField>
+		void copy(const PolynomialMatrix<OtherField, PMType::polfirst> & M, size_t beg, size_t end, size_t start=0){
+			//cout<<"copying.....polfirst to polfirst.....other field"<<endl;
+			Hom<OtherField,Field> hom(M.field(),field()) ;
+			for (size_t k=0;k<_row*_col;k++){
+				size_t j=0;
+				for (size_t i=beg;i<=end;i++,j++){
+					hom.image(ref(k,start+j),M.get(k,i));
+				}
+			}
+		}
+
+                // copy elt from M[beg..end], _size must be >= end-beg+1
+		// M is stored with the hybrid format matrowfirst
+		template<typename OtherField>
+		void copy(const PolynomialMatrix<OtherField, PMType::matrowfirst> & M, size_t beg, size_t end, size_t start=0){
+			//std::cout<<"copying.....matrowfirst to polfirst.....same field"<<std::endl;
+			Hom<OtherField,Field> hom(M.field(),field()) ;
+			const size_t ls = COPY_BLOCKSIZE;
+            for(size_t k=0;k<_row;k++){
+                for (size_t i = beg; i <= end; i+=ls)
+                    for (size_t j = 0; j < _col; j+=ls)        
+                        for (size_t _i = i; _i < std::min(end+1, i + ls); _i++)
+                            for (size_t _j = j; _j < std::min(_col, j + ls);++_j)
+                                hom.image(ref(k,_j,start+_i-beg), M.get(k,_j,_i));
+            }
+        }
+
+	
 
 		template<typename Mat>
 		void copy(const Mat& M){
@@ -250,14 +286,21 @@ namespace LinBox{
 			copy(M,0,M.size()-1);
 		}
 
+
+        template<class MatPoly>
+        void copy(const SubPolynomialMatrix<MatPoly>& M, size_t beg, size_t end, size_t start=0){
+            copy(*M._ptr, beg+M._shift, end+M._shift,start);
+        }
+
+        
 		// rebind functor to change base field (e.g. apply modulo reduction)
 		template<typename _Tp1>
 		struct rebind {
-			typedef PolynomialMatrix<PMType::polfirst,PMStorage::plain, Field> Self_t;
-			typedef PolynomialMatrix<PMType::polfirst,PMStorage::plain, _Tp1> Other_t;
+			typedef PolynomialMatrix<Field, PMType::polfirst> Self_t;
+			typedef PolynomialMatrix<_Tp1, PMType::polfirst> Other_t;
 
 			void operator() (Other_t& Ap,
-					 const Self_t&  A){
+                             const Self_t&  A){
 				Hom<Field, _Tp1> hom(A.field(), Ap.field()) ;
 				for (size_t i = 0; i < A._row * A._col; i++)
 					for (size_t j = 0; j < A._size; j++)
@@ -265,26 +308,11 @@ namespace LinBox{
 			}
 		};
 
-		// get access to the the k-th coeff  of the ith matrix entry
-		inline Element& ref(size_t i, size_t k){
-			return _rep[i*_store+k];
-		}
-		inline Element& ref(size_t i, size_t j, size_t k){
-			return ref(i*_col+j,k);
-		}
-		inline Element get(size_t i, size_t k)const {
-			return _rep[i*_store+k];
-		}
-		inline Element get(size_t i, size_t j, size_t k) const{
-			return get(i*_col+j,k);
-		}
 
-		size_t rowdim()  const {return _row;}
-		size_t coldim()  const {return _col;}
-		size_t degree()  const {return _size-1;}
-		size_t size()    const {return _size;}
-		size_t storage() const {return _store;}
-		const Field& field()  const {return *_fld;}
+    
+        size_t realmeminfo()const {			
+			return _rep.size()*element_storage(field());
+		}
 
 		void dump(std::ostream& os) const {
 			os<<_row<<" "<<_col<<" "<<_size<<std::endl;
@@ -297,142 +325,159 @@ namespace LinBox{
 		
 		std::ostream& write(std::ostream& os) const { return write(os,0,_size-1);}
 
-                std::ostream& write(std::ostream& os, size_t deg_min, size_t deg_max) const {
-                        integer c;
-                        int wid=-1,b;
-                        field().cardinality (c);
+        std::ostream& write(std::ostream& os, size_t deg_min, size_t deg_max) const {
+            integer c;
+            int wid=-1,b;
+            field().cardinality (c);
 
-                        if (c >0){
-                                wid = (int) ceil (log ((double) c) / M_LN10);
-                                b= ((int) ceil (log ((double) _size) / M_LN10));
-                                wid*=10*(b*(b-1)/2.);
-                        }
-                        std::cout<<"Matrix([";
-                        for (size_t i = 0; i< _row;++i) {
-                            os << "  [ ";
-                            for (size_t j = 0;j<_col;++j){
-                                os.width (wid);
-                                field().write(os,get(i,j,deg_min));
-                                for (size_t k=deg_min+1;k<deg_max;++k){
-                                    os<<"+";
-                                    field().write(os,get(i,j,k))<<"*x^"<<k-deg_min;
-                                }
-                                if (deg_max-deg_min>0){
-                                    os<<"+";
-                                    field().write(os,get(i,j,deg_max))<<"*x^"<<deg_max-deg_min;
-                                }
-                                os<<(j<_col-1?",":"" );
-                            }
-                            os << (i<_row-1?"],":"]" )<< std::endl;
-                        }
-			std::cout<<"]);";
-		
-                	return os;
+            if (c >0){
+                wid = (int) ceil (log ((double) c) / M_LN10);
+                b= ((int) ceil (log ((double) _size) / M_LN10));
+                wid*=10*(b*(b-1)/2.);
+            }
+            os<<"Matrix([";
+            for (size_t i = 0; i< _row;++i) {
+                os << "  [ ";
+                for (size_t j = 0;j<_col;++j){
+                    os.width (wid);
+                    field().write(os,get(i,j,deg_min));
+                    for (size_t k=deg_min+1;k<deg_max;++k){
+                        os<<"+";
+                        field().write(os,get(i,j,k))<<"*x^"<<k-deg_min;
+                    }
+                    if (deg_max-deg_min>0){
+                        os<<"+";
+                        field().write(os,get(i,j,deg_max))<<"*x^"<<deg_max-deg_min;
+                    }
+                    os<<(j<_col-1?",":"" );
                 }
-
-		Element* getPointer(){return _rep.data();}
-		const Element* getPointer() const {return _rep.data();}
-
-		size_t realmeminfo()const {
-			return _row*_col*(_store*element_storage(field())+sizeof(Polynomial));}
-		// return _rep.capacity()*sizeof(Element)+_repview.capacity()*sizeof(Polynomial);}
-	
-		size_t meminfo()const { return _rep.size()*sizeof(Element);}
-
-		void changeField(const Field& F){_fld=&F;}
+                os << (i<_row-1?"],":"]" );//<< std::endl;
+            }
+			os<<"]);";
 		
-	private:
-		size_t           _store;
-		std::vector<Polynomial> _repview;
-		//std::vector<Element>    _rep;
-		VECT _rep;
-		size_t             _row;
-		size_t             _col;
-		size_t            _size;
-		const Field*       _fld;
+            return os;
+        }
+
+        size_t rowdim()  const {return _row;}
+		size_t coldim()  const {return _col;}
+		size_t degree()  const {return _size-1;}
+		size_t size()    const {return _size;}
+        size_t poly_stride() const {return 1;}
+		const Field& field()  const {return _rep.field();}        
+
+
+    private:
+        Data     _rep;
+        size_t   _row;
+		size_t   _col;
+		size_t  _size;
 	};
 
-	// Class for Polynomial Matrix stored as a Polynomial of Matrices
-	template<class _Field>
-	class PolynomialMatrix<PMType::matfirst,PMStorage::plain,_Field> {
-	public:
-		typedef _Field Field;
-		typedef typename Field::Element   Element;
-		typedef BlasMatrix<Field>         Matrix;
-		typedef std::vector<Element>       Polynomial;
-		typedef PolynomialMatrix<PMType::matfirst,PMStorage::plain,_Field>  Self_t;
-		typedef PolynomialMatrix<PMType::polfirst,PMStorage::plain,_Field> Other_t;		
-			       
-		typedef Matrix value_type;
+    // Class for Polynomial Matrix stored as a Polynomial of Matrices
+    template<class _Field>
+	class PolynomialMatrix<_Field, PMType::matfirst> {
+    public:
+        typedef _Field                         Field;
+		typedef typename Field::Element      Element;
+		typedef std::vector<Element,AlignedAllocator<Element, Alignment::DEFAULT>> VECT;
+        typedef BlasVector<Field>                             Data;        
 
-		PolynomialMatrix() {}
+        typedef BlasSubmatrix<BlasMatrix<Field>>                    Matrix;
+        typedef BlasSubmatrix<const BlasMatrix<Field>>         constMatrix;
+        typedef BlasSubvector<Data>                             Polynomial;
+        typedef const BlasSubvector<const Data>            constPolynomial;        
 
-		PolynomialMatrix(const Self_t&) = delete;
+		typedef PolynomialMatrix<_Field, PMType::matfirst>  Self_t;
+        typedef SubPolynomialMatrix<Self_t>                   view;
+        typedef SubPolynomialMatrix<const Self_t>       const_view;
+        
+		// construct a polynomial matrix in f[x]^(m x n) of degree (s-1)
+        PolynomialMatrix(const Field& f, size_t r, size_t c, size_t s)
+            : _rep(f,r*c*s), _row(r), _col(c), _size(s) {}
+
+
+        /*******************************
+         * Data access functionnalities*
+         ******************************/
+        
+        // retrieve the polynomial at entry (i,j) in the matrix
+		Polynomial       operator()(size_t i, size_t j)      {return operator()(i*_col+j);}
+		constPolynomial  operator()(size_t i, size_t j)const {return operator()(i*_col+j);}
+
+        // retrieve the polynomial at the position i in the storage of the matrix
+		Polynomial       operator()(size_t i)      {return      Polynomial(_rep, i,_row*_col,_size);}
+		constPolynomial  operator()(size_t i)const {return constPolynomial(_rep, i,_row*_col,_size);}
+        
+		// get write access to the the k-th coeff  of the ith matrix entry
+		Element& ref(size_t i, size_t k)      {return _rep[i+k*_row*_col];}
+        // get read access to the the k-th coeff  of the ith matrix entry
+		Element  get(size_t i, size_t k) const {return _rep[i+k*_row*_col];}
+        // get write access to the the k-th coeff  of the entry (i,j) in matrix 
+		Element& ref(size_t i, size_t j, size_t k)      { return _rep[i*_col+j+k*_row*_col];}
+        // get read access to the the k-th coeff  of the entry (i,j) in matrix 
+		Element  get(size_t i, size_t j, size_t k) const { return _rep[i*_col+j+k*_row*_col];}
+
+		Element*       getPointer()       {return _rep.getPointer();}
+		const Element* getPointer() const {return _rep.getConstPointer();}
+
 		
-		PolynomialMatrix(const Field& f, size_t r, size_t c, size_t s) :
-			_rep(s,Matrix(f)), _row(r), _col(c), _size(s), _fld(&f) {
-			//_row(r), _col(c), _size(s), _fld(&f) {			
-			for(size_t i=0;i<s;i++)
-				_rep[i].init(r,c);
-			//integer p;
-			//std::cout<<"(ALLOC) matfirst at "<<this<<" : "<<r<<"x"<<c<<" - size= "<<s<<" ==> "<<MB(realmeminfo())<<" Mo"<<std::endl;
-			ADD_MEM(realmeminfo());
-		}
+        // copy the matrix of degree k into A
+        template<typename DenseMatrix>
+		DenseMatrix& getMatrix(DenseMatrix& A, size_t k) const {
+			return A= Matrix(field(), getPointer()+k*_row*_col, _row,_col,_col);            
+        }
 
-		~PolynomialMatrix(){
-			DEL_MEM(realmeminfo());
-			//std::cout<<"(FREE) matfirst at "<<this<<" : "<<_row<<"x"<<_col<<" - size= "<<_size<<" ==> "<<MB(realmeminfo())<<" Mo"<<std::endl;
-			//integer p;
-			//std::cout<<"PMatrix Desallocating : "<<_row*_col*_size*length(_fld->characteristic(p))/1000000.<<"Mo"<<std::endl;
-		}
+        // copy the matrix A into the matrix of degree
+        template<typename DenseMatrix>
+		void setMatrix(const DenseMatrix& A, size_t k)  {
+            if (A.getPointer() != getPointer()+k*_row*_col){
+                //std::cout<<"NEW MAtPol setMatrix: "<<A.getPointer()<<"<>"<<getPointer()+k*_row*_col<<std::endl;
+                auto it=A.Begin();
+                for(size_t i=0;i<_row*_col;i++,it++)
+                    ref(i,k)=*it ;
+            }
+            else {
+                //std::cout<<"NEW MAtPol setMatrix: does nothing\n";
+            }
+		} 
 
-		void clear(){
-			_rep.resize(0,Matrix(*_fld));
-		}
 
-		
-		// retrieve the matrix of degree k in the polynomial matrix
-		inline Matrix&       operator[](size_t k)      {return _rep[k];}
-		inline const Matrix& operator[](size_t k)const {return _rep[k];}
+        Matrix       operator[](size_t k)          {return Matrix(field(), getPointer()+k*_row*_col, _row,_col,_col);}
+		constMatrix  operator[](size_t k) const {return constMatrix(field(), getPointer()+k*_row*_col, _row,_col,_col);}
 
-		// retrieve the polynomial at entry (i,j) in the matrix
-		inline Polynomial  operator()(size_t i, size_t j) const {
-			Polynomial A(_size);
-			for(size_t k=0;k<_size;k++)
-				A[k]=_rep[k].getEntry(i,j);
-			return A;
-		}
-		// retrieve the polynomial at psotion (i) in the storage of the matrix
-		inline Polynomial  operator()(size_t i) const {
-			Polynomial A(_size);
-			for(size_t k=0;k<_size;k++)
-				A[k]=*(_rep[k].Begin()+i);
-			return A;
-		}
+        view       at(size_t i, size_t j)       {return view(*this,i,j);}
+        const_view at(size_t i, size_t j) const {return const_view(*this,i,j);}
+
 
 		// resize the polynomial length of the polynomial matrix
 		void resize(size_t s){
-			_rep.resize(s,Matrix(*_fld));
-			if (s>_size){
-				for(size_t i=_size;i<s;i++)
-					_rep[i].init(_row,_col);
-			}
-			_size=s;
+			_rep.resize(s*_row*_col,field().zero);
+            _size=s;
 		}
-
-		void setsize(size_t s){resize(s);}
 		
 		// copy elt from M[beg..end], _size must be >= j-i
-		template <size_t storage>
-		void copy(const PolynomialMatrix<PMType::matfirst,storage,Field>& M, size_t beg, size_t end, size_t start=0){
+		void copy(const PolynomialMatrix<Field, PMType::matfirst>& M, size_t beg, size_t end, size_t start=0){
 			//cout<<"copying.....matfirst to matfirst.....same field"<<endl;
-			for (size_t i=beg;i<=end;i++)
-				_rep[start+i-beg]=M[i];
-
+            FFLAS::fassign(field(), (end-beg+1)*_row*_col, M.getPointer()+beg*_row*_col,1, getPointer()+start,1);
 		}
 
-		template<size_t storage, typename OtherField>
-		void copy(const PolynomialMatrix<PMType::matfirst,storage,OtherField> & M, size_t beg, size_t end, size_t start=0){
+        // copy elt from M[beg..end], _size must be >= end-beg+1
+		// M is stored with the hybrid format matrowfirst
+		void copy(const PolynomialMatrix<Field, PMType::matrowfirst>& M, size_t beg, size_t end, size_t start=0){
+			//std::cout<<"copying.....matrowfirst to polfirst.....same field"<<std::endl;
+			const size_t ls = COPY_BLOCKSIZE;
+                for (size_t i = beg; i <= end; i+=ls)
+                    for (size_t j = 0; j < _row; j+=ls)        
+                        for (size_t _i = i; _i < std::min(end+1, i + ls); _i++)
+                            for (size_t _j = j; _j < std::min(_row, j + ls);++_j)
+                                for(size_t k=0;k<_col;k++)
+                                    ref(k,_j,start+_i-beg)= M.get(k,_j,_i);            
+        }
+
+
+        
+		template<typename OtherField>
+		void copy(const PolynomialMatrix<OtherField, PMType::matfirst> & M, size_t beg, size_t end, size_t start=0){
 			//cout<<"copying.....matfirst to matfirst.....other field"<<endl;
 			Hom<OtherField,Field> hom(M.field(),field()) ;
 			for (size_t i=beg;i<=end;i++)
@@ -442,7 +487,7 @@ namespace LinBox{
 
 		// copy elt from M[beg..end], _size must be >= end-beg+1
 		// M is stored as a Matrix of Polynomials
-		void copy(const Other_t& M, size_t beg, size_t end, size_t start=0){
+		void copy(const PolynomialMatrix<Field, PMType::polfirst>& M, size_t beg, size_t end, size_t start=0){
 			//cout<<"copying.....polfirst to matfirst.....same field"<<endl;
 			const size_t ls = COPY_BLOCKSIZE; // Loop unrooling to be more cache-friendly (cache block transposition)
 			for(size_t i = beg; i < end+1; i+=ls)
@@ -456,7 +501,7 @@ namespace LinBox{
 		// copy elt from M[beg..end], _size must be >= end-beg+1
 		// M is stored as a Matrix of Polynomials with a different field
 		template<typename OtherField>
-		void copy(const PolynomialMatrix<PMType::polfirst,PMStorage::plain,OtherField> & M, size_t beg, size_t end, size_t start=0){
+		void copy(const PolynomialMatrix<OtherField, PMType::polfirst> & M, size_t beg, size_t end, size_t start=0){
 			//cout<<"copying.....polfirst to matfirst.....other field"<<endl;
 			const size_t ls = COPY_BLOCKSIZE;
 			Hom<OtherField,Field> hom(M.field(),field()) ;
@@ -474,15 +519,22 @@ namespace LinBox{
 			copy(M,0,M.size()-1);
 		}
 
+
+        template<class MatPoly>
+        void copy(const SubPolynomialMatrix<MatPoly>& M, size_t beg, size_t end, size_t start=0){
+            copy(*M._ptr, beg+M._shift, end+M._shift,start);
+        }
+
+        
+        
 		// rebind functor to change base field (e.g. apply modulo reduction)
 		template<typename _Tp1>
 		struct rebind {
-			typedef PolynomialMatrix<PMType::matfirst, PMStorage::plain,Field>  Self_t;
-			typedef PolynomialMatrix<PMType::matfirst, PMStorage::plain,_Tp1>  Other_t;
+			typedef PolynomialMatrix<Field, PMType::matfirst>  Self_t;
+			typedef PolynomialMatrix<_Tp1, PMType::matfirst>  Other_t;
 
-			template<size_t storage>
-			void operator() (PolynomialMatrix<PMType::matfirst, storage,_Tp1>& Ap,
-					 const PolynomialMatrix<PMType::matfirst, storage,Field>&  A){
+			void operator() (PolynomialMatrix<_Tp1, PMType::matfirst>& Ap,
+                             const PolynomialMatrix<Field, PMType::matfirst>&  A){
 				Hom<Field, _Tp1> hom(A.field(), Ap.field()) ;
 				for (size_t j = 0; j < A._size; j++)
 					for (size_t i = 0; i < A._row * A._col; i++)
@@ -490,43 +542,19 @@ namespace LinBox{
 			}
 		};
 
-		typedef PolynomialMatrix<PMType::matfirst,PMStorage::plain,Field>           plain;
-		typedef PolynomialMatrix<PMType::matfirst,PMStorage::view,Field>             view;
-		typedef PolynomialMatrix<PMType::matfirst,PMStorage::const_view,Field> const_view;
-
-		inline view       at(size_t i, size_t j)      {return view(*this,i,j);}
-		inline const_view at(size_t i, size_t j)const {return const_view(*this,i,j);}
-
-
-		// get access to the the k-th coeff  of the ith matrix entry
-		inline Element& ref(size_t i, size_t k){
-			//return 	_rep[k].refRep()[i];
-			return *(_rep[k].getPointer()+i); 
-		}
-		inline Element& ref(size_t i, size_t j, size_t k){
-			return ref(i*_col+j,k);
-		}
-		inline Element get(size_t i, size_t k)const {
-			return *(_rep[k].Begin()+i);
-		}
-		inline Element get(size_t i, size_t j, size_t k) const{
-			return get(i*_col+j,k);
-		}
-
-		inline size_t rowdim() const {return _row;}
-		inline size_t coldim() const {return _col;}
-		inline size_t degree() const {return _size-1;}
-		inline size_t real_degree() const {
+       
+        size_t real_degree() const {
 			MatrixDomain<Field> MD(field());
 			size_t d= _size-1;
-			while(d>0 && MD.isZero(_rep[d])) d--;
+			while(d>0 && MD.isZero(operator[](d))) d--;
 			return d;
 		}
-		inline size_t size()   const {return _size;}
-		inline size_t storage()const {return _size;}
-		
-		inline const Field& field()  const {return *_fld;}
 
+        size_t realmeminfo()const {			
+			return _rep.size()*element_storage(field());
+		}
+
+        
 		void dump(std::ostream& os) const {
 			os<<_row<<" "<<_col<<" "<<_size<<std::endl;
 			for(size_t k=0;k<_size;k++)
@@ -538,215 +566,453 @@ namespace LinBox{
 		
 		std::ostream& write(std::ostream& os) const { return write(os,0,real_degree());}
 
-                std::ostream& write(std::ostream& os, size_t deg_min, size_t deg_max) const {
-                        integer c;
-                        int wid,b;
-                        field().cardinality (c);
+        std::ostream& write(std::ostream& os, size_t deg_min, size_t deg_max) const {
+            integer c;
+            int wid,b;
+            field().cardinality (c);
 
-                        if (c >0)
-                                wid = (int) ceil (log ((double) c) / M_LN10);
+            if (c >0)
+                wid = (int) ceil (log ((double) c) / M_LN10);
 			else
-				wid=(int) ceil (log ((double) _rep[0].getEntry(0,0)) / M_LN10);
+				wid=(int) ceil (log ((double) get(0,0,0)) / M_LN10);
 
 			b= ((int) ceil (log ((double) (deg_max-deg_min+1)) / M_LN10));
 			wid*=10*(b*(b-1)/2.);
-			std::cout<<"Matrix([";
+			os<<"Matrix([";
 			for (size_t i = 0; i< _row;++i) {
-                                os << "  [ ";
-                                for (size_t j = 0;j<_col;++j){
-                                        os.width (wid);
-                                        field().write(os,_rep[deg_min].getEntry(i,j));
-                                        for (size_t k=deg_min+1;k<deg_max;++k){
-                                            if (!field().isZero(_rep[k].getEntry(i,j))){
-                                                os<<"+";
-                                                field().write(os,_rep[k].getEntry(i,j))<<"*x^"<<k-deg_min;
-                                            }
-                                        }
-                                        if (deg_max-deg_min>0){
-                                            if (!field().isZero(_rep[deg_max].getEntry(i,j))){
-                                                os<<"+";
-                                                field().write(os,_rep[deg_max].getEntry(i,j))<<"*x^"<<deg_max-deg_min;
-                                            }
-                                        }
-					os<<(j<_col-1?",":"" );
-                                }
-                                os << (i<_row-1?"],":"]" )<< std::endl;
+                os << "  [ ";
+                for (size_t j = 0;j<_col;++j){
+                    os.width (wid);
+                    //field().write(os,_rep[deg_min].getEntry(i,j));
+                    field().write(os,get(i,j,deg_min));
+                    for (size_t k=deg_min+1;k<deg_max;++k){
+                        if (!field().isZero(get(i,j,k))){
+                            os<<"+";
+                            field().write(os,get(i,j,k))<<"*x^"<<k-deg_min;
                         }
-			std::cout<<"]);";
-			return os;
+                    }
+                    if (deg_max-deg_min>0){
+                        if (!field().isZero(get(i,j,deg_max))){
+                            os<<"+";
+                            field().write(os,get(i,j,deg_max))<<"*x^"<<deg_max-deg_min;
+                        }
+                    }
+					os<<(j<_col-1?",":"" );
                 }
+                os << (i<_row-1?"],":"]" );//<< std::endl;
+            }
+			os<<"]);";
+			return os;
+        }        	
 
-		size_t realmeminfo()const {
-			
-			return _size*(_row*_col*element_storage(field())+sizeof(Matrix));
-		}
-		
-		// NEED FOR YUHASZ
-		typedef typename std::vector<Matrix>::const_iterator const_iterator;
-		const_iterator begin() const {return _rep.begin();}
+        size_t rowdim()  const {return _row;}
+		size_t coldim()  const {return _col;}
+		size_t degree()  const {return _size-1;}
+		size_t size()    const {return _size;}
+        size_t poly_stride() const {return _row*_col;}        
+		const Field& field()  const {return _rep.field();}        
 
-	private:
-		std::vector<Matrix>     _rep;
-		size_t             _row;
-		size_t             _col;
-		size_t            _size;
-		const Field*       _fld;
+    private:
+        Data     _rep;
+        size_t   _row;
+		size_t   _col;
+		size_t  _size;
 	};
 
-	template<typename _Field, size_t T, size_t S>
-	std::ostream& operator<<(std::ostream& os, const PolynomialMatrix<T,S,_Field>& P) {
-		return P.write(os);
-	}
 
-	// Class to handle the view of a Polynomial Matrix according to some degree range
-	// the view is read/write
-	template<class _Field>
-	class PolynomialMatrix<PMType::matfirst,PMStorage::view,_Field> {
-	public:
-		typedef _Field Field;
-		typedef typename Field::Element   Element;
-		typedef BlasMatrix<Field>          Matrix;
-		typedef std::vector<Element>        Polynomial;
+    // Class for Polynomial Matrix stored with an hybrid variant that store each row as a Polynomial of row 
+    template<class _Field>
+	class PolynomialMatrix<_Field, PMType::matrowfirst> {
+    public:
+        typedef _Field                         Field;
+		typedef typename Field::Element      Element;
+		typedef std::vector<Element,AlignedAllocator<Element, Alignment::DEFAULT>> VECT;
+        typedef BlasVector<Field>                             Data;        
 
-		PolynomialMatrix() {}
+        typedef BlasSubmatrix<BlasMatrix<Field>>                    Matrix;
+        typedef BlasSubmatrix<const BlasMatrix<Field>>         constMatrix;
+        typedef BlasSubvector<Data>                             Polynomial;
+        typedef const BlasSubvector<const Data>            constPolynomial;        
 
-		// constructor of a view between i and j from a plain Polynomial Matrix
-		PolynomialMatrix(PolynomialMatrix<PMType::matfirst,PMStorage::plain,Field>& M, size_t i,size_t j)
-			: _ptr(&M), _size(j-i+1), _shift(i)
-		{//cout<<i<<"-"<<j<<" -> "<<M.size()<<endl;
-			linbox_check(i<M.size() && i<=j && j< M.size());}
-
-		// constructor of a view between i and j from a view Polynomial Matrix
-		PolynomialMatrix(PolynomialMatrix<PMType::matfirst,PMStorage::view,Field>& M, size_t i,size_t j)
-			: _ptr(M._ptr), _size(j-i+1), _shift(i+M._shift)
-		{//cout<<i<<"-"<<j<<" -> "<<M.size()<<endl;
-			linbox_check(i<M.size() && i<=j && j< M.size());}
+		typedef PolynomialMatrix<_Field, PMType::matrowfirst>  Self_t;
+        typedef SubPolynomialMatrix<Self_t>                      view;
+        typedef SubPolynomialMatrix<const Self_t>          const_view;
+        
+		// construct a polynomial matrix in f[x]^(m x n) of degree (s-1)
+        PolynomialMatrix(const Field& f, size_t r, size_t c, size_t s)
+            : _rep(f,r*c*s), _row(r), _col(c), _size(s), _stride(c*s) {}
 
 
+        /*******************************
+         * Data access functionnalities*
+         ******************************/
+        
+        // retrieve the polynomial at entry (i,j) in the matrix
+		Polynomial       operator()(size_t i, size_t j)      {return operator()(i*_stride+j);}
+		constPolynomial  operator()(size_t i, size_t j)const {return operator()(i*_stride+j);}
+
+        // retrieve the polynomial at the position i in the storage of the matrix
+		Polynomial       operator()(size_t i)      {return      Polynomial(_rep, i,_row,_size);}
+		constPolynomial  operator()(size_t i)const {return constPolynomial(_rep, i,_row,_size);}
+        
+		// get write access to the the k-th coeff  of the ith matrix entry
+		Element& ref(size_t i, size_t k)      {return ref(i/_col, i%_col, k);}
+        // get read access to the the k-th coeff  of the ith matrix entry
+		Element  get(size_t i, size_t k) const {return get(i/_col, i%_col, k);}
+        // get write access to the the k-th coeff  of the entry (i,j) in matrix 
+		Element& ref(size_t i, size_t j, size_t k)      { return _rep[i*_stride+j+k*_col];}
+        // get read access to the the k-th coeff  of the entry (i,j) in matrix 
+		Element  get(size_t i, size_t j, size_t k) const{ return _rep[i*_stride+j+k*_col];}
+
+		Element*       getPointer()       {return _rep.getPointer();}
+		const Element* getPointer() const {return _rep.getConstPointer();}
+
+		
 		// retrieve the matrix of degree k in the polynomial matrix
-		inline Matrix&       operator[](size_t k)      {return (*_ptr)[k+_shift];}
-		inline const Matrix& operator[](size_t k)const {return (*_ptr)[k+_shift];}
+		//Matrix       operator[](size_t k)       {return      Matrix(field(), getPointer()+k*_row*_col, _row,_col,_col);}
 
-		// retrieve the polynomial at entry (i,j) in the matrix
-		Polynomial    operator()(size_t i, size_t j){
-			Polynomial A(_size, Matrix(_ptr->field()));
-			for(size_t k=0;k<_size;k++)
-				A[k]=(*_ptr)[k+_shift].refEntry(i,j);
-			return A;
+        // copy the matrix of degree k into A
+        template<typename DenseMatrix>
+		DenseMatrix& getMatrix(DenseMatrix& A, size_t k) const {
+			return A= Matrix(field(), getPointer()+k*_col, _row,_col,_stride);            
+        }
+
+        // copy the matrix A into the matrix of degree
+        template<typename DenseMatrix>
+		void setMatrix(const DenseMatrix& A, size_t k)  {
+            if (A.getPointer() != getPointer()+k*_col){
+                //std::cout<<"NEW MAtPol setMatrix: "<<A.getPointer()<<"<>"<<getPointer()+k*_row*_col<<std::endl;
+                auto it=A.Begin();
+                for(size_t i=0;i<_row*_col;i++,it++)
+                    ref(i,k)=*it ;
+            }
+            else {
+                //std::cout<<"NEW MAtPol setMatrix: does nothing\n";
+            }
+		} 
+
+        //Matrix&&  operator[](size_t k)          {Matrix tmp(field(), getPointer()+k*_row*_col, _row,_col,_col); return std::move(tmp);}
+        Matrix       operator[](size_t k)          {return Matrix(field(), getPointer()+k*_col, _row,_col,_stride);}
+		constMatrix  operator[](size_t k) const {return constMatrix(field(), getPointer()+k*_col, _row,_col,_stride);}
+
+        view       at(size_t i, size_t j)       {return view(*this,i,j);}
+        const_view at(size_t i, size_t j) const {return const_view(*this,i,j);}
+
+
+
+
+		// resize the polynomial length of the polynomial matrix
+		void resize(size_t s){
+			if (s==_size) return;
+			if (s>_size ){
+				_rep.resize(s*_row*_col, field().zero);
+
+                if (s*_col > _stride) { // move data if needed
+                    for(size_t i=0;i<_row;i++){
+                        size_t k1=(i+1)*_stride-1;
+                        size_t k2=(i+1)*s*_col-1;
+                        for(size_t j=0;j<_size*_col;k1--,k2--)
+                            _rep[k2]=_rep[k1];
+                    }
+                }
+                // put zeros at the right place
+                size_t diff=(s-_size)*_col;
+                for(size_t i=0;i<_row;i++){
+                    size_t k1=(i+1)*_stride;
+                    FFLAS::fzero(field(), diff, getPointer()+k1, 1);
+                }
+                // correct the stride when necessary
+                if (s*_col > _stride) _stride=s*_col;
+			}
+			// if (s< _size) do nothing since we have the right stride
+			_size=s;
+		}	
+
+/*
+ * PG -> NEXT LINES NEED TO BE CHECKED AND TESTED
+ *
+ */
+
+        // copy elt from M[beg..end], _size must be >= end_beg+1
+		void copy(const PolynomialMatrix<Field, PMType::matfirst>& M, size_t beg, size_t end, size_t start=0){
+            // 1st version, write access compliant
+            for (size_t i=0;i<_row;i++)
+                FFLAS::fassign(field(), (end-beg+1),_col, M.getPointer()+beg*_row*_col+i*_col, _row*_col, getPointer()+start+i*_stride,_col);
+            // 2nd version, read access compliant
+            // for (size_t i=beg;i<end;i++)
+            //     FFLAS::fassign(field(), _row, _col, M.getPointer()+beg*_row*_col, _col, getPointer()+start+(beg-i),_stride);
+            
+		}        
+		
+		// copy elt from M[beg..end], _size must be >= end-beg+1
+		void copy(const PolynomialMatrix<Field, PMType::matrowfirst>& M, size_t beg, size_t end, size_t start=0){
+            FFLAS::fassign(field(), _row, (end-beg+1)*_col, M.getPointer()+beg*_col, M._stride, getPointer()+start,_stride);
+		}
+        
+		// copy elt from M[beg..end], _size must be >= end-beg+1
+		// M is stored as a Matrix of Polynomials
+		void copy(const PolynomialMatrix<Field, PMType::polfirst>& M, size_t beg, size_t end, size_t start=0){
+			const size_t ls = COPY_BLOCKSIZE; // Loop unrooling to be more cache-friendly (cache block transposition)
+            for(size_t k=0;k<_row;k++) {
+                for(size_t i = beg; i < end+1; i+=ls)
+                    for (size_t j = 0; j < _col; j+=ls)
+                        for (size_t _i = i; _i < std::min(end+1, i + ls); _i++) {
+                            for (size_t _j = j; _j < std::min(_col, j + ls); _j++)
+                                ref(k,_j,start+_i-beg) = M.get(k,_j,_i);
+                        }
+            }
+        }
+
+		// copy elt from M[beg..end], _size must be >= end-beg+1
+		// M is stored as a Matrix of Polynomials with a different field
+        // PG -> could be specialized for each storage
+		template<typename OtherField, PMType Storage>
+		void copy(const PolynomialMatrix<OtherField, Storage> & M, size_t beg, size_t end, size_t start=0){
+			Hom<OtherField,Field> hom(M.field(),field()) ;
+			for (size_t i=beg;i<=end;i++)
+				for (size_t j=0;j<_row*_col;j++)
+					hom.image(ref(j,start+i-beg) , M.get(j,i));
 		}
 
-		// get access to the the k-th coeff  of the ith matrix entry
-		inline Element& ref(size_t i, size_t k){
-			return 	_ptr->ref(i,k+_shift);
-		}
-		inline Element& ref(size_t i, size_t j, size_t k){
-			return ref(i*coldim()+j,k);
-		}
-		inline Element get(size_t i, size_t k)const {
-			return 	_ptr->get(i,k+_shift);
-		}
-		inline Element get(size_t i, size_t j, size_t k) const{
-			return get(i*coldim()+j,k);
-		}
 
-		inline size_t rowdim() const {return _ptr->rowdim();}
-		inline size_t coldim() const {return _ptr->coldim();}
-		inline size_t degree() const {return _size-1;}
-		inline size_t size()   const {return _size;}
-		inline const Field& field()  const {return _ptr->field();}
-
-		typedef PolynomialMatrix<PMType::matfirst,PMStorage::plain,Field>           plain;
-		typedef PolynomialMatrix<PMType::matfirst,PMStorage::view,Field>             view;
-		typedef PolynomialMatrix<PMType::matfirst,PMStorage::const_view,Field> const_view;
 
 		template<typename Mat>
 		void copy(const Mat& M){
-			this->copy(M,0,M.size()-1);
+			linbox_check (M.size()==_size && M.rowdim()== _row && M.coldim()== _col);
+			copy(M,0,M.size()-1);
 		}
 
-		template<typename Mat>
-		void copy(const Mat& M, size_t beg, size_t end){_ptr->copy(M,beg,end,_shift);}
+
+        template<class MatPoly>
+        void copy(const SubPolynomialMatrix<MatPoly>& M, size_t beg, size_t end, size_t start=0){
+            copy(*M._ptr, beg+M._shift, end+M._shift,start);
+        }
+
+        
+        
+		// rebind functor to change base field (e.g. apply modulo reduction)
+		template<typename _Tp1>
+		struct rebind {
+              typedef PolynomialMatrix<Field, PMType::matrowfirst>  Self_t;
+              typedef PolynomialMatrix<_Tp1, PMType::matrowfirst>  Other_t;
+              
+              void operator() (PolynomialMatrix<_Tp1, PMType::matrowfirst>& Ap,
+                               const PolynomialMatrix<Field, PMType::matrowfirst>&  A){
+                  Hom<Field, _Tp1> hom(A.field(), Ap.field()) ;
+                  for (size_t j = 0; j < A._size; j++)
+                      for (size_t i = 0; i < A._row * A._col; i++)
+                          hom.image (Ap.ref(i,j), A.get(i,j));
+              }
+        };
 
 
-		inline view       at(size_t i, size_t j)      {return view(*this,i,j);}
-		inline const_view at(size_t i, size_t j)const {return const_view(*this,i,j);}
+        size_t real_degree() const {
+			MatrixDomain<Field> MD(field());
+			size_t d= _size-1;
+			while(d>0 && MD.isZero(operator[](d))) d--;
+			return d;
+		}
 
-		std::ostream& write(std::ostream& os) const { return _ptr->write(os,_shift,_shift+_size-1);}
+		size_t realmeminfo()const {			
+			return _rep.size()*element_storage(field());
+		}
 
-		friend class  PolynomialMatrix<PMType::matfirst,PMStorage::const_view,Field>;
+        
+		void dump(std::ostream& os) const {
+			os<<_row<<" "<<_col<<" "<<_size<<std::endl;
+			for(size_t k=0;k<_row;k++)
+				for(size_t i=0;i<_size*_col;i++)		
+					os<<getPointer()[k*_stride+i]<<" ";
+			os<<std::endl;
+		}
 
-	private:
-		PolynomialMatrix<PMType::matfirst,PMStorage::plain,Field>* _ptr;
+		
+		std::ostream& write(std::ostream& os) const { return write(os,0,real_degree());}
+
+        std::ostream& write(std::ostream& os, size_t deg_min, size_t deg_max) const {
+            integer c;
+            int wid,b;
+            field().cardinality (c);
+
+            if (c >0)
+                wid = (int) ceil (log ((double) c) / M_LN10);
+			else
+				wid=(int) ceil (log ((double) get(0,0,0)) / M_LN10);
+
+			b= ((int) ceil (log ((double) (deg_max-deg_min+1)) / M_LN10));
+			wid*=10*(b*(b-1)/2.);
+			os<<"Matrix([";
+			for (size_t i = 0; i< _row;++i) {
+                os << "  [ ";
+                for (size_t j = 0;j<_col;++j){
+                    os.width (wid);
+                    //field().write(os,_rep[deg_min].getEntry(i,j));
+                    field().write(os,get(i,j,deg_min));
+                    for (size_t k=deg_min+1;k<deg_max;++k){
+                        if (!field().isZero(get(i,j,k))){
+                            os<<"+";
+                            field().write(os,get(i,j,k))<<"*x^"<<k-deg_min;
+                        }
+                    }
+                    if (deg_max-deg_min>0){
+                        if (!field().isZero(get(i,j,deg_max))){
+                            os<<"+";
+                            field().write(os,get(i,j,deg_max))<<"*x^"<<deg_max-deg_min;
+                        }
+                    }
+					os<<(j<_col-1?",":"" );
+                }
+                os << (i<_row-1?"],":"]" );//<< std::endl;
+            }
+			os<<"]);";
+			return os;
+        }        	
+
+        size_t rowdim()  const {return _row;}
+		size_t coldim()  const {return _col;}
+		size_t degree()  const {return _size-1;}
+		size_t size()    const {return _size;}
+        size_t poly_stride() const {return _row;}        
+		const Field& field()  const {return _rep.field();}        
+
+    private:
+        Data     _rep;
+        size_t   _row;
+		size_t   _col;
 		size_t  _size;
-		size_t _shift;
+        size_t _stride;
 	};
 
-	// Class to handle the view of a Polynomial Matrix according to some degree range
-	// the view is read only
-	template<class _Field>
-	class PolynomialMatrix<PMType::matfirst,PMStorage::const_view,_Field> {
-	public:
-		typedef _Field Field;
-		typedef typename Field::Element   Element;
-		typedef BlasMatrix<Field>          Matrix;
-		typedef std::vector<Element>        Polynomial;
 
-		PolynomialMatrix() {}
+    
+	template<typename _Field, LinBox::PMType T>
+	std::ostream& operator<<(std::ostream& os, const PolynomialMatrix<_Field,T>& P) {
+		return P.write(os);
+	}
+    
+
+
+    template<typename MatPoly>
+    struct SubPolyConst{
+        typedef typename MatPoly::Matrix         Matrix;
+        typedef typename MatPoly::constMatrix    constMatrix;
+        typedef typename MatPoly::view             view;
+        typedef typename MatPoly::const_view const_view;                
+    };
+
+    template<typename MatPoly>
+    struct SubPolyConst <const MatPoly>{
+        typedef typename MatPoly::constMatrix    constMatrix;
+        typedef typename MatPoly::constMatrix         Matrix;
+        typedef typename MatPoly::const_view        view;
+        typedef typename MatPoly::const_view const_view;                
+    };
+
+    
+    
+	// Class to handle the view of a Polynomial Matrix according to some degree range
+	template<class MatPoly>
+	class SubPolynomialMatrix {
+	public:
+        typedef SubPolynomialMatrix<MatPoly>     Self_t;
+        typedef typename MatPoly::Field           Field;
+        typedef typename MatPoly::Element       Element;
+        typedef typename MatPoly::Polynomial Polynomial;        
+        typedef typename SubPolyConst<MatPoly>::Matrix            Matrix;
+        typedef typename SubPolyConst<MatPoly>::constMatrix  constMatrix;
+        typedef typename SubPolyConst<MatPoly>::view             view;
+        typedef typename SubPolyConst<MatPoly>::const_view const_view;
+        
+		SubPolynomialMatrix() {}
 
 		// constructor of a view between i and j from a plain Polynomial Matrix
-		PolynomialMatrix(const PolynomialMatrix<PMType::matfirst,PMStorage::plain,Field>& M, size_t i,size_t j)
+		SubPolynomialMatrix(MatPoly& M, size_t i,size_t j)
 			: _ptr(&M), _size(j-i+1), _shift(i)
-		{linbox_check(i<M.size() && i<=j && j< M.size());}
+		{
+            linbox_check(i<M.size() && i<=j && j< M.size());
+            //if (i>=M.size() || (i>j)) {_size=0;}
+        }
 
-		// constructor of a view between i and j from a view Polynomial Matrix
-		PolynomialMatrix(const PolynomialMatrix<PMType::matfirst,PMStorage::view,Field>& M, size_t i,size_t j)
+		// constructor of a view between i and j from a Sub Polynomial Matrix
+		SubPolynomialMatrix(Self_t & M, size_t i,size_t j)
 			: _ptr(M._ptr), _size(j-i+1), _shift(i+M._shift)
-		{linbox_check(i<M.size() && i<=j && j< M.size());}
+		{
+            linbox_check(i<M.size() && i<=j && j< M.size());
+            //if (i>=M.size() || (i>j)) {_size=0;}
+        }
 
+         // copy the matrix of degree k into A
+        template<typename DenseMatrix>
+		DenseMatrix& getMatrix(DenseMatrix& A, size_t k) const {
+            return _ptr->getMatrix(A,k+_shift);
+        }
 
-		// constructor of a view between i and j from a view Polynomial Matrix
-		PolynomialMatrix(const PolynomialMatrix<PMType::matfirst,PMStorage::const_view,Field>& M, size_t i,size_t j)
-			: _ptr(M._ptr), _size(j-i+1), _shift(i+M._shift)
-		{//cout<<"("<<i<<","<<j<<") -> "<<M.size()<<endl;
-			linbox_check(i<M.size()&& i<=j && j<= M.size());}
+        // copy the matrix A into the matrix of degree 
+        template<typename DenseMatrix>
+		void setMatrix(const DenseMatrix& A, size_t k)  {
+            return _ptr->setMatrix(A,k+_shift);
+        }
 
+        // only shrink down the size, increase is not possible
+		void resize(size_t s){
+            if (s<=_size) _size=s;
+            else{
+                std::cerr<<"LinBox -> SubPolynomialMatrix: resizing with larger size than available, exiting ..."<<std::endl;
+                throw LinboxError("SubPolynomialMatrix: resizing with larger size than available, exiting ...");
+            }
+        }
 
+        template <typename Mat>
+        void copy(const Mat & M, size_t beg, size_t end){
+            _ptr->copy(M,beg,end,_shift);
+		}
+        
+		template<typename Mat>
+        void copy(const Mat& M){
+            copy(M,0,M.size()-1);
+        }
+
+        
 		// retrieve the matrix of degree k in the polynomial matrix
-		inline const Matrix& operator[](size_t k)const {return (*_ptr)[k+_shift];}
+        Matrix  operator[](size_t k) const {return _ptr->operator[](k+_shift);}
 
 		// retrieve the polynomial at entry (i,j) in the matrix
-		inline Polynomial    operator()(size_t i, size_t j){
-			Polynomial A(_size, Matrix(_ptr->field()));
-			for(size_t k=0;k<_size;k++)
-				A[k]=(*_ptr)[k+_shift].refEntry(i,j);
-			return A;
-		}
+        Polynomial    operator()(size_t i, size_t j){
+            return Polynomial(_ptr->operator()(i,j),i, _ptr->poly_stride(), j-i+1);
+        }
 
-		inline Element get(size_t i, size_t k) const {
-			return 	_ptr->get(i,k+_shift);
-		}
-		inline Element get(size_t i, size_t j, size_t k) const{
-			return get(i*coldim()+j,k);
-		}
+        Element get(size_t i, size_t k) const { return 	_ptr->get(i,k+_shift);}
+
+        Element get(size_t i, size_t j, size_t k) const{ return get(i*coldim()+j,k);}
+
+        size_t rowdim() const {return _ptr->rowdim();}
+        size_t coldim() const {return _ptr->coldim();}
+        size_t degree() const {return _size-1;}
+        size_t size()   const {return _size;}
+        const Field& field()  const {return _ptr->field();}
 
 
-		inline size_t rowdim() const {return _ptr->rowdim();}
-		inline size_t coldim() const {return _ptr->coldim();}
-		inline size_t degree() const {return _size-1;}
-		inline size_t size()   const {return _size;}
-		inline const Field& field()  const {return _ptr->field();}
+        const_view at(size_t i, size_t j) const {return const_view(*_ptr,i+_shift,j+_shift);}
+        view       at(size_t i, size_t j)       {return       view(*_ptr,i+_shift,j+_shift);}
+        
+		std::ostream& write(std::ostream& os) const { if (_size!=0) return _ptr->write(os,_shift,_shift+_size-1); else return os;}
 
-		typedef PolynomialMatrix<PMType::matfirst,PMStorage::plain,Field>           plain;
-		typedef PolynomialMatrix<PMType::matfirst,PMStorage::const_view,Field> const_view;
+        
 
-		inline const_view at(size_t i, size_t j)const {return const_view(*this,i,j);}
-		std::ostream& write(std::ostream& os) const { return _ptr->write(os,_shift,_shift+_size-1);}
+        // give friendship access to the main PolynomialMatrix classes
+        template<class Field, enum PMType>
+        friend class PolynomialMatrix;
 
-	private:
-		const PolynomialMatrix<PMType::matfirst,PMStorage::plain,Field>* _ptr;
+	protected:
+		MatPoly* _ptr;
 		size_t  _size;
 		size_t _shift;
 	};
+
+	template<typename MatPoly>      
+	std::ostream& operator<<(std::ostream& os, const SubPolynomialMatrix<MatPoly>& P) {
+		return P.write(os);
+	}
 
 
 
@@ -755,8 +1021,6 @@ namespace LinBox{
 } //end of namespace LinBox
 
 #endif // __LINBOX_polynomial_matrix_H
-
-#endif //__OLD_POLYNOMIAL_MATRIX
 
 // Local Variables:
 // mode: C++
