+++ tests/test-matpoly-mult.C
@@ -74,15 +74,19 @@ void randomMat (Rand& r, Mat& m) {
 
 
 template<typename Field, typename Rand>
-void randomMatPol(Rand& r,  PolynomialMatrix<PMType::matfirst,PMStorage::plain,Field>& A){
-	for(size_t i=0;i<A.size();i++)
-		randomMat(r, A[i]);
+void randomMatPol(Rand& r,  PolynomialMatrix<Field,PMType::matfirst>& A){
+	for(size_t i=0;i<A.size();i++){
+        auto Ai=A[i];
+        randomMat(r, Ai);
+    }
 }
 
 template<typename Field, typename Rand>
-void randomMatPol(Rand& r,  PolynomialMatrix<PMType::polfirst,PMStorage::plain,Field>& A){
-	for(size_t i=0;i<A.rowdim()*A.coldim();i++)
-		randomVect(r, A(i));
+void randomMatPol(Rand& r,  PolynomialMatrix<Field,PMType::polfirst>& A){
+	for(size_t i=0;i<A.rowdim()*A.coldim();i++){
+        auto Ai=A(i);
+        randomVect(r, Ai);
+    }
 }
 
 
@@ -177,14 +181,14 @@ bool launchTest(const Field& F, size_t n, uint64_t b, long d, long seed){
     bool ok=true;
     Givaro::Integer samplesize(1); samplesize<<=b;
     typename Field::RandIter G(F,seed,samplesize);
-	typedef PolynomialMatrix<PMType::polfirst,PMStorage::plain,Field> MatrixP;
+	typedef PolynomialMatrix<Field,PMType::polfirst> MatrixP;
 	ostream& report = LinBox::commentator().report();
 	report<<"Polynomial matrix (polfirst) testing over ";F.write(report)<<std::endl;
 	ok&=check_matpol_mul<MatrixP> (F,G,n,d);
 	ok&=check_matpol_midp<MatrixP> (F,G,n,d);
 	ok&=check_matpol_midpgen<MatrixP> (F,G,n,d);
 
-	//typedef PolynomialMatrix<PMType::matfirst,PMStorage::plain,Field> PMatrix;
+	//typedef PolynomialMatrix<Field,PMType::matfirst> PMatrix;
 	// std::cerr<<"Polynomial matrix (matfirst) testing:\n";F.write(std::cerr)<<std::endl;
 	// check_matpol_mul<PMatrix> (F,G,n,d);
 	// check_matpol_midp<PMatrix> (F,G,n,d);
@@ -223,7 +227,7 @@ bool runTest(uint64_t n, uint64_t d, long seed){
         report<<"prime bits : "<<p.bitsize()<<std::endl;
 		Field F((int32_t)p);
 		ok&=launchTest (F,n,bits,d,seed);
-        commentator().stop(MSG_STATUS (ok), (const char *) 0,"Half wordsize generic prime");
+        commentator().stop(MSG_STATUS (ok), (const char *) 0,"Half wordsize normal prime");
 	}
 
 	// multi-precision prime
